pyproject.toml
[project]
name = "ace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-openai",
    "langchain-google-genai",
    "langchain-community",
    "ollama",
    "python-dotenv",
    "pydantic",
    "numpy",
    "pandas",
    "scikit-learn",
    "deepeval",
    "ragas",
    "pytest",
    "black",
    "flake8",
    "sphinx",
    "mkdocs",
    "streamlit",
    "sentence-transformers",
    "torch",
    "pypdf",
    "chromadb",
    "sentencepiece",
]

[tool.setuptools.packages.find]
include = ["ace_framework*"]
